set background images to pattern and sulci columns in order to distinguish them

+++ python/morphologist/sulpated/sulpated_data.py
@@ -353,11 +353,13 @@ class SulcalPattern(object):
         return osp.exists(self.lock_file)
 
     def lock(self):
-        with open(self.lock_file, 'w') as f:
-            print(datetime.datetime.now(), file=f)
+        if osp.exists(osp.dirname(self.lock_file)):
+            with open(self.lock_file, 'w') as f:
+                print(datetime.datetime.now(), file=f)
 
     def unlock(self):
-        os.unlink(self.lock_file)
+        if osp.exists(osp.dirname(self.lock_file)):
+            os.unlink(self.lock_file)
 
 
 class FileLock(object):
